+++ Array.js
@@ -1,20 +1,10 @@
 "use strict";
 
-
-
-/**
- * Creates array with duplicates of the items of this array.
- * This function works recursively and will call .duplicate() for the
- * items that implement this function.
- * @def function Array.duplicate ()
- * @return Array
- * @author Borislav Peev <borislav.asdf@gmail.com>
- */
 Object.defineProperty( Array.prototype, 'duplicate', {
 	value: function () {
 		var ret = [].concat( this );
 		for ( var i = ret.length - 1; i >= 0; --i ) {
-			var r = ret[i];
+			var r = ret[ i ];
 			if ( r instanceof Object && r.duplicate instanceof Function ) {
 				ret[i] = r.duplicate();
 			}
@@ -24,43 +14,41 @@ Object.defineProperty( Array.prototype, 'duplicate', {
 	writable: true
 } );
 
-
-
-/**
-Alias for {@see Array.concat()}, just for consistency.
-@def function Array.merge ( ... )
-*/
 Object.defineProperty( Array.prototype, 'merge', {
 	value: Array.prototype.concat,
 	writable: true
 } );
 
-/**
- * Retrieves or sets the last element of the array.
- * @def var Array.last
- * @var any|undefined Returns undefined if attempting to get the last element of zero-length array.
- * @author Borislav Peev <borislav.asdf@gmail.com>
- */
 Object.defineProperty( Array.prototype, 'last', {
 	get: function () {
 		var i = this.length - 1;
-		return i >= 0 ? this[i] : undefined;
+		return i >= 0 ? this[ i ] : undefined;
 	},
 	set: function ( v ) {
 		var i = this.length - 1;
-		return i >= 0 ? this[i] = v : this[0] = v;
+		return i >= 0 ? this[ i ] = v : this[ 0 ] = v;
 	}
 } );
 
 
-Object.defineProperty( Array.prototype, 'contains', { value: function ( value ) {
+Object.defineProperty( Array.prototype, 'contains', {
+	value: function ( value ) {
 		return this.indexOf( value ) >= 0;
 	},
-	writable: false
+	writable: true
 } );
 
 
-Object.defineProperty( Array.prototype, 'unique', { value: function ( value ) {
+Object.defineProperty( Array.prototype, 'containsEx', {
+	value: function ( value ) {
+		return this.indexOfEx( value ) >= 0;
+	},
+	writable: true
+} );
+
+
+Object.defineProperty( Array.prototype, 'unique', {
+	value: function () {
 		var unique = [];
 		for ( var item of this ) {
 			if ( unique.indexOf( item ) < 0 ) {
@@ -69,5 +57,35 @@ Object.defineProperty( Array.prototype, 'unique', { value: function ( value ) {
 		}
 		return unique;
 	},
-	writable: false
+	writable: true
+} );
+
+Object.defineProperty( Array.prototype, 'indexOfEx', {
+	value: function ( callback, offset ) {
+		if ( !(callback instanceof Function) ) {
+			return this.indexOf( callback, offset || 0 );
+		}
+		for ( var i = offset || 0, iend = this.length; i < iend; ++i ) {
+			if ( callback( this[ i ], i, this ) ) {
+				return i;
+			}
+		}
+		return -1;
+	},
+	writable: true
+} );
+
+Object.defineProperty( Array.prototype, 'lastIndexOfEx', {
+	value: function ( callback, offset ) {
+		if ( !(callback instanceof Function) ) {
+			return this.lastIndexOf( callback, offset !== undefined ? offset : this.length - 1 );
+		}
+		for ( var i = offset || this.length - 1; i >= 0; --i ) {
+			if ( callback( this[ i ], i, this ) ) {
+				return i;
+			}
+		}
+		return -1;
+	},
+	writable: true
 } );