+++ Prototype.js
@@ -10,3 +10,4 @@ if ( typeof window != "undefined" && typeof require == "undefined" ) {
 /*@*/require( './Function.js' );
 /*@*/require( './Number.js' );
 /*@*/require( './Boolean.js' );
+/*@*/require( './RegExp.js' );